Use modules
Seemed like a waste to set up my tsconfig and package.json and then
just throw all code into a single file. Don't know if this organization
is good or not, but at least it's better than what was there. Also
standardized naming a bit, using 'make...' for my factory functions and
'I...' for types and interfaces.