Inline sources in emitted JavaScript
This way, for GitHub Pages, I don't have to actually include the
source TypeScript files. I also experimented with copying over the
/src directory to the build folder and then using the tsconfig option
sourcesRoot. That also worked as far as I could tell, but this way is
a bit neater and it doesn't require making the commands for 'npm run
build' longer.